fix components that error in 2.4+
