+++ addon/mixins/components/inputable.js
@@ -9,6 +9,20 @@ export default Ember.Mixin.create({
 
   hasError: Ember.computed.alias('errors.length'),
 
+  hasValue: Ember.computed('value', function() {
+    const value = this.get('value');
+
+    return value ? true : false;
+  }),
+
+  showError: Ember.computed('hasError', 'required', 'hasValue', function() {
+    const hasError = this.get('hasError');
+    const required = this.get('required');
+    const hasValue = this.get('hasValue');
+
+    return hasError || (required && !hasValue);
+  }),
+
   inputId: Ember.computed(function() {
     return `bootstrap-component-${generateUUID()}`;
   }),
