+++ addon/templates/components/bootstrap-input.hbs
@@ -1,40 +1,25 @@
-<div class="{{concat-class-names classNames}} form-group bootstrap-input-component {{if hasSuccess 'has-success'}} {{if hasWarning 'has-warning'}} {{if (or hasError (and required (not hasValue))) 'has-error'}}">
-  {{#if hasBlock}}
-    {{yield inputId}}
-  {{else}}
-    {{#if label}}
-      <label for="{{inputId}}" class="control-label {{if customLabelCss customLabelCss ''}} {{if srOnly 'sr-only' ''}}">{{label}}</label>
-    {{/if}}
-    {{#if isChecked}}
-      {{input
-        checked=value
-        elementId=inputId
-        type=type
-        placeholder=placeholder
-        readonly=readonly
-        disabled=disabled
-        required=required}}
-    {{else}}
-      {{input
-        tabindex=tabindex
-        value=value
-        elementId=inputId
-        type=type
-        placeholder=placeholder
-        classNames="form-control"
-        readonly=readonly
-        disabled=disabled
-        key-press=keyPress
-        key-up=keyUp
-        key-down=keyDown
-        required=required}}
-    {{/if}}
-    {{#if errors}}
-      <div class="form-control-static alert alert-danger" role="alert">
-        {{#each errors as |error|}}
-          <span>{{error.message}}</span>
-        {{/each}}
-      </div>
-    {{/if}}
+{{#if hasBlock}}
+  {{yield inputId}}
+{{else}}
+  <label for="{{inputId}}" class="control-label {{if customLabelCss customLabelCss ''}} {{if srOnly 'sr-only' ''}}">{{label}}</label>
+  {{input
+    tabindex=tabindex
+    value=value
+    elementId=inputId
+    type=type
+    placeholder=placeholder
+    classNames="form-control"
+    readonly=readonly
+    disabled=disabled
+    key-press=keyPress
+    key-up=keyUp
+    key-down=keyDown
+    required=required}}
+  {{#if errors}}
+    <div class="form-control-static alert alert-danger" role="alert">
+      {{#each errors as |error|}}
+        <span>{{error.message}}</span>
+      {{/each}}
+    </div>
   {{/if}}
-</div>
+{{/if}}
