+++ addon/templates/components/bootstrap-phone-mask-input.hbs
@@ -1,31 +1,29 @@
-<div class="{{concat-class-names classNames}} form-group bootstrap-input-component {{if hasSuccess 'has-success'}} {{if hasWarning 'has-warning'}} {{if (or hasError (and required (not hasValue))) 'has-error'}}">
-  {{#if hasBlock}}
-    {{yield inputId}}
-  {{else}}
-    {{#if label}}
-      <label for="{{inputId}}" class="control-label {{if customLabelCss customLabelCss ''}} {{if srOnly 'sr-only' ''}}">{{label}}</label>
-    {{/if}}
-    {{masked-input
-      mask=phoneMask
-      tabindex=tabindex
-      value=value
-      elementId=inputId
-      classNames="form-control"
-      type=type
-      readonly=readonly
-      disabled=disabled
-      placeholder=placeholder
-      placeholderChar=placeholderChar
-      key-press=keyPress
-      key-up=keyUp
-      key-down=keyDown
-      required=required}}
-    {{#if errors}}
-      <div class="form-control-static alert alert-danger" role="alert">
-        {{#each errors as |error|}}
-          <span>{{error.message}}</span>
-        {{/each}}
-      </div>
-    {{/if}}
+{{#if hasBlock}}
+  {{yield inputId}}
+{{else}}
+  {{#if label}}
+    <label for="{{inputId}}" class="control-label {{if customLabelCss customLabelCss ''}} {{if srOnly 'sr-only' ''}}">{{label}}</label>
   {{/if}}
-</div>
+  {{masked-input
+    mask=phoneMask
+    tabindex=tabindex
+    value=value
+    elementId=inputId
+    classNames="form-control"
+    type=type
+    readonly=readonly
+    disabled=disabled
+    placeholder=placeholder
+    placeholderChar=placeholderChar
+    key-press=keyPress
+    key-up=keyUp
+    key-down=keyDown
+    required=required}}
+  {{#if errors}}
+    <div class="form-control-static alert alert-danger" role="alert">
+      {{#each errors as |error|}}
+        <span>{{error.message}}</span>
+      {{/each}}
+    </div>
+  {{/if}}
+{{/if}}
