+++ addon/templates/components/bootstrap-textarea.hbs
@@ -1,26 +1,24 @@
-<div class="{{concat-class-names classNames}} form-group bootstrap-textarea-component {{if hasSuccess 'has-success'}} {{if hasWarning 'has-warning'}} {{if (or hasError (and required (not hasValue))) 'has-error'}}">
-  {{#if hasBlock}}
-    {{yield inputId}}
-  {{else}}
-    {{#if label}}
-      <label for="{{inputId}}" class="control-label {{if srOnly 'sr-only' ''}}">{{label}}</label>
-    {{/if}}
-    {{textarea
-      tabindex=tabindex
-      value=value
-      elementId=inputId
-      placeholder=placeholder
-      classNames="form-control"
-      rows=rows
-      cols=cols
-      disabled=disabled
-      required=required}}
-    {{#if errors}}
-      <div class="alert alert-danger" role="alert">
-        {{#each errors as |error|}}
-          <span>{{error.message}}</span>
-        {{/each}}
-      </div>
-    {{/if}}
+{{#if hasBlock}}
+  {{yield inputId}}
+{{else}}
+  {{#if label}}
+    <label for="{{inputId}}" class="control-label {{if srOnly 'sr-only' ''}}">{{label}}</label>
   {{/if}}
-</div>
+  {{textarea
+    tabindex=tabindex
+    value=value
+    elementId=inputId
+    placeholder=placeholder
+    classNames="form-control"
+    rows=rows
+    cols=cols
+    disabled=disabled
+    required=required}}
+  {{#if errors}}
+    <div class="alert alert-danger" role="alert">
+      {{#each errors as |error|}}
+        <span>{{error.message}}</span>
+      {{/each}}
+    </div>
+  {{/if}}
+{{/if}}
